Fix sun shield max temperatures with DRE
There was already a fix that was supposed to prevent this. Even with
that fix, however, I was still seeing DRE's maximum temperatures being
far below what they should have been. This is a more involved fix, but
it makes the DRE temperatures for the heat shields match what is
expected in all cases.

This ended up being somewhat more involved than I thought it would. You
can set `depletedMaxTemp` on the `ModuleHeatShield` but it doesn't
actually apply unless you also set `skinMaxOperationalTemp` and
`maxOperationalTemp` on `ModuleAeroReentry`.

Normally you wouldn't have to worry about this. Unfortunately, the
default values for those fields are far below what is configured for the
sun shields and DRE will only ever adjust those fields down. As such,
the second part of this patch sets the initial values there to the
appropriate starting values.